Split common provider code into a base class.
