+++ lib/puppet/provider/rustup_internal/shell.rb
@@ -1,51 +1,26 @@
 # frozen_string_literal: true
 
-Puppet::Type.type(:rustup_internal).provide(:shell) do
+require_relative "../rustup_exec"
+
+Puppet::Type.type(:rustup_internal).provide(
+  :shell, :parent => Puppet::Provider::RustupExec
+) do
   desc "Run shell-based `rustup` installer on UNIX-like platforms."
 
-  # It’s not really reliable to query rustup installations by user, since they
-  # could be installed with CARGO_HOME anywhere.
-  def self.instances
-    fail "Cannot query rustup installations."
-  end
-
-  # For whatever reason, the resource expected this. An alternate solution is
-  # to call mk_resource_methods, but that adds a bunch more unnecessary methods
-  # that modify @property_hash. We don’t use @property_hash at all, since we
-  # can just access the resource itself.
-  def ensure=(value)
-  end
-
-  def cargo_home
-    resource[:cargo_home] || File.join(home_path(), ".cargo")
-  end
-
-  def rustup_home
-    resource[:rustup_home] || File.join(home_path(), ".rustup")
-  end
-
+  # Determine if `rustup` has been installed on the system for this user
   def exists?
     # FIXME? this actually checks that root can execute the file. Also, it
     # doesn’t check that it’s not a directory.
     File.executable?(rustup_path())
   end
 
-  # Changes have been made to the resource; apply them.
-  def flush
-    if exists?
-      if resource[:ensure] == :absent
-        uninstall
-      elsif resource[:ensure] == :latest
-        update
-      end
-    elsif resource[:ensure] != :absent
-      # Does not exist, but should.
-      install
-    end
-  end
+protected
 
-private
-
+  # Install `rustup` for the first time.
+  #
+  # Will only be called if both:
+  #   * exists? == false
+  #   * ensure != :absent
   def install
     # Puppet::Util::Execution.execute can’t accept an IO stream or a string as
     # stdin, so we save the script as a file and pipe it into stdin. (We don’t
@@ -75,36 +50,25 @@ private
     end
   end
 
+  # Update previously installed `rustup`.
+  #
+  # Will only be called if both:
+  #   * exists? == true
+  #   * ensure == :latest
   def update
     rustup "self", "update"
   end
 
+  # Uninstall previously installed `rustup`.
+  #
+  # Will only be called if both:
+  #   * exists? == true
+  #   * ensure == :absent
   def uninstall
     rustup "self", "uninstall", "-y"
   end
 
-  def execute(command, stdin_file: nil, raise_on_failure: true)
-    environment = {
-      "PATH" => path_env(),
-      "RUSTUP_HOME" => rustup_home(),
-      "CARGO_HOME" => cargo_home(),
-    }
-
-    stdin_message = stdin_file ? " and stdin_file #{stdin_file}" : ""
-    debug("Running #{command} for user #{resource[:user]} with environment #{environment}#{stdin_message}")
-
-    # FIXME timeout?
-    Puppet::Util::Execution.execute(
-      command,
-      :failonfail => raise_on_failure,
-      :uid => resource[:user],
-      :combine => true,
-      :stdinfile => stdin_file,
-      :override_locale => false,
-      :custom_environment => environment,
-    )
-  end
-
+  # Download a URL into a stream
   def download_into(url, output)
     client = Puppet.runtime[:http]
     client.get(url, options: {include_system_store: true}) do |response|
@@ -120,31 +84,4 @@ private
       end
     end
   end
-
-  # Convenience function to run rustup
-  def rustup(*args)
-    execute([rustup_path()] + args)
-  end
-
-  def rustup_path
-    File.join(bin_path(), "rustup")
-  end
-
-  def path_env
-    "#{bin_path()}:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin"
-  end
-
-  def bin_path
-    File.join(cargo_home(), "bin")
-  end
-
-  def home_path
-    Etc.getpwnam(resource[:user]).dir
-  end
-
-  def debug(message)
-    Puppet.debug("rustup_internal: #{message}")
-  end
 end
-
-
